[#13] feat: 버튼 필터링 기능 구현
버튼 형태로 지역이름을 이용해 특정 데이터를 하이라이트 하는 방식으로 필터링 기능을 구현했습니다.
Bar Chart는 backgroundColor와 borderColor를 이용해 하이라이트 처리했습니다.
Area Chart는 borderColor를 이용해 하이라이트 처리했습니다.
