+++ src/Router.tsx
@@ -1,3 +1,14 @@
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+
+import { ChartPage } from './pages';
+
 export default function Router() {
-  return;
+  const router = createBrowserRouter([
+    {
+      path: '/',
+      element: <ChartPage />,
+    },
+  ]);
+
+  return <RouterProvider router={router} />;
 }
